Created sections for the landing page
