+++ _daily_logs/group_logs.md
@@ -773,4 +773,6 @@ export default config;
   - Copied over basic query for `plantsBasicInfo` list. Hard coded the params above the return statement. Will need to fix that later when implementing queries within other components
   
 
-## Monday Dec 18th
+## Wednesday Dec 19th
+
+-
